requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
awscli==1.22.34
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.4
botocore==1.23.34
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cohere==4.57
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
dbus-python==1.2.18
Deprecated==1.2.14
discord==2.3.2
discord.py==2.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
effdet==0.4.1
emoji==2.10.1
exceptiongroup==1.2.2
fastapi==0.115.0
fastavro==1.9.7
ffmpy==0.4.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iopath==0.1.10
jeepney==0.7.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonpath-python==1.0.6
keyring==23.5.0
kiwisolver==1.4.5
langdetect==1.0.9
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
opencv-python==4.8.0.76
orjson==3.10.7
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
protobuf==4.23.4
pyasn1==0.4.8
pycocotools==2.0.7
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.0.9
pypdf==4.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
rich==13.9.1
roman==3.3
rsa==4.8
ruff==0.6.8
s3transfer==0.5.0
safetensors==0.3.2
scipy==1.10.1
SecretStorage==3.3.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.38.6
sympy==1.12
systemd-python==234
tabulate==0.9.0
timm==0.9.12
tokenizers==0.15.2
tomlkit==0.12.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.1
triton==2.2.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.3
uvicorn==0.31.0
wadllib==1.3.6
websockets==12.0
wrapt==1.16.0
XlsxWriter==3.1.9
yarl==1.13.1
zipp==3.17.0